requirements_bot.txt
discord.py>=2.3.0
aiohttp>=3.8.0
asyncio-python>=0.1.0